Fix [https://trac.osgeo.org/mapguide/ticket/291].
Update DrawString to use doubles for the position of the string instead of integers.  The use
of doubles allows two things:
* subpixel positioning accuracy for renderers which support AA text rendering
* support for device spaces which are not pixel-based


git-svn-id: http://svn.osgeo.org/mapguide/trunk@2666 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ MgDev/Common/Renderers/AGGW2DRewriter.cpp
@@ -716,7 +716,7 @@ WT_Result agr_process_text (WT_Text & text, WT_File & file)
         wchar_t* uni_text = WT_String::to_wchar(text.string().length(), text.string().unicode());
 
         //width is not used by AGGRenderer - can leave at zero
-        AGGRenderer::DrawString((agg_context*)rewriter->GetW2DTargetImage(), uni_text, (int)dstpts->x_coord(0), (int)dstpts->y_coord(0), 0.0, hgt, rsfont, color, angleRad);
+        AGGRenderer::DrawString((agg_context*)rewriter->GetW2DTargetImage(), uni_text, dstpts->x_coord(0), dstpts->y_coord(0), 0.0, hgt, rsfont, color, angleRad);
 
         delete [] uni_text;
