+++ MgDev/Common/Stylization/RS_FontEngine.h
@@ -101,8 +101,8 @@ public:
                                                float*           offsets) = 0;
 
     STYLIZATION_API virtual void DrawString(const RS_String& s,
-                                            int              x,
-                                            int              y,
+                                            double           x,
+                                            double           y,
                                             double           width,
                                             double           height,
                                             const RS_Font*   font,
